+++ heima-leadnews/heima-leadnews-service/heima-leadnews-content/src/main/java/com/heima/content/mapper/tip/ApArticleTipOrderMapper.java
@@ -1,0 +1,9 @@
+package com.heima.content.mapper.tip;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.heima.model.article.pojos.ApArticleTipOrder;
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface ApArticleTipOrderMapper extends BaseMapper<ApArticleTipOrder> {
+}